Moves node priority calculation into NodeContractor.
